Common experiment code.
- most of the code of main experiment was moved to "common" package
- logic responsible for validation and topology was moved to separated
packages
- moved QoS workload class to common
- update tendencies
- fixed error propagation and handling
