+++ experiments/memcached-sensitivity-profile/topology/core_selection_test.go
@@ -1,4 +1,4 @@
-package main
+package topology
 
 import (
 	"testing"
